seperated core logic and fixed installation guide(#90)
